+++ src/menu/menu.h
@@ -138,11 +138,11 @@ class DListMenuDescriptor : public DMenuDescriptor
 public:
 	TArray<DMenuItemBase *> mItems;
 	int mSelectedItem;
-	int mSelectOfsX;
-	int mSelectOfsY;
+	double mSelectOfsX;
+	double mSelectOfsY;
 	FTextureID mSelector;
 	int mDisplayTop;
-	int mXpos, mYpos;
+	double mXpos, mYpos;
 	int mWLeft, mWRight;
 	int mLinespacing;	// needs to be stored for dynamically created menus
 	int mAutoselect;	// this can only be set by internal menu creation functions
@@ -288,7 +288,7 @@ class DMenuItemBase : public DObject
 {
 	DECLARE_CLASS(DMenuItemBase, DObject)
 public:
-	int mXpos, mYpos;
+	double mXpos, mYpos;
 	FNameNoInit mAction;
 	bool mEnabled;
 
@@ -298,7 +298,7 @@ public:
 	bool SetValue(int i, int value);
 	bool GetValue(int i, int *pvalue);
 	void OffsetPositionY(int ydelta) { mYpos += ydelta; }
-	int GetY() { return mYpos; }
+	double GetY() { return mYpos; }
 };	
 
 //=============================================================================
@@ -355,7 +355,7 @@ DMenuItemBase * CreateOptionMenuItemStaticText(const char *name, bool v);
 DMenuItemBase * CreateOptionMenuItemSubmenu(const char *label, FName cmd, int center);
 DMenuItemBase * CreateOptionMenuItemControl(const char *label, FName cmd, FKeyBindings *bindings);
 DMenuItemBase * CreateOptionMenuItemJoyConfigMenu(const char *label, IJoystickConfig *joy);
-DMenuItemBase * CreateListMenuItemPatch(int x, int y, int height, int hotkey, FTextureID tex, FName command, int param);
-DMenuItemBase * CreateListMenuItemText(int x, int y, int height, int hotkey, const char *text, FFont *font, PalEntry color1, PalEntry color2, FName command, int param);
+DMenuItemBase * CreateListMenuItemPatch(double x, double y, int height, int hotkey, FTextureID tex, FName command, int param);
+DMenuItemBase * CreateListMenuItemText(double x, double y, int height, int hotkey, const char *text, FFont *font, PalEntry color1, PalEntry color2, FName command, int param);
 
 #endif
